pyproject.toml
[project]
name = "travel_map"
version = "0.0.1"
requires-python=">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "osmnx>=2.0.0b3",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.13.1",
    "uvicorn>=0.34.3",
    "scikit-learn>=1.7.0",
    "gpxpy>=1.6.2",
    "pytest-xdist>=3.8.0",
    "httpx>=0.28.1",
    "redis>=6.4.0",
    "fastapi-cache2>=0.2.2",
    "aiolimiter>=1.2.1",
    "psutil>=7.0.0",
    "sentry-sdk[fastapi]>=2.38.0",
    "sqlmodel>=0.0.25",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "prometheus-client>=0.23.1",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "pydantic[email]>=2.11.6",
    "sqladmin[full]>=0.21.0",
    "psycopg2-binary>=2.9.10",
    "fastapi-mail>=1.5.0",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.2.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "debugpy>=1.8.14",
    "factory-boy>=3.3.3",
    "folium>=0.20.0",
    "html2image>=2.0.7",
    "imgkit>=1.2.3",
    "matplotlib>=3.10.3",
    "playwright>=1.55.0",
    "pyqt5>=5.15.11",
    "pytest-asyncio>=1.2.0",
    "pytest-coverage>=0.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "ty>=0.0.1a9",
    "vulture>=2.14",
]

[tool.pytest.ini_options]
addopts = [
    "-v",
    # "-s",
    "--cov=app"
]

[tool.uv]
package = true
